requirements.txt
torch
torchvision
transformers
datasets
huggingface-hub
pillow
seqeval
torchmetrics
evaluate
pytesseract

